+++ src/onegov.activity/onegov/activity/upgrade.py
@@ -3,10 +3,10 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 
-from onegov.activity import Booking, Period
+from onegov.activity import Booking, Period, Occasion
 from onegov.core.orm.types import UUID, JSON
 from onegov.core.upgrade import upgrade_task
-from sqlalchemy import Boolean, Column, Enum, ForeignKey
+from sqlalchemy import Boolean, Column, Enum, ForeignKey, Integer
 from sqlalchemy.orm import joinedload
 
 
@@ -85,3 +85,17 @@ def add_data_column_to_period(context):
 
     context.session.flush()
     context.operations.alter_column('periods', 'data', nullable=False)
+
+
+@upgrade_task('Add attendee_count column to occasion')
+def add_attendee_count_column_to_occasion(context):
+    context.operations.add_column('occasions', Column(
+        'attendee_count', Integer, nullable=True, default=0
+    ))
+
+    for occasion in context.session.query(Occasion):
+        occasion.attendee_count = len(occasion.accepted)
+
+    context.session.flush()
+    context.operations.alter_column(
+        'occasions', 'attendee_count', nullable=False)